Changed size related variable to type size_t

+++ inc/libbase/spi_master.h
@@ -21,7 +21,7 @@ public:
 
   SPIMaster(Name, uint32_t baud);
   ~SPIMaster();
-  void Exchange(uint8_t*, uint8_t*, uint32_t len);
+  void Exchange(uint8_t*, uint8_t*, size_t len);
 
 private:
   SPI_Type* spin;
